requirements.txt
python-telegram-bot==21.9
python-dotenv==1.1.0
gspread==6.1.4
google-auth==2.38.0
APScheduler==3.11.0
httpx==0.28.1
flask==3.1.0
waitress==3.0.1
